requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.1
django-cors-headers==3.4.0
djangorestframework==3.11.1
drf-yasg==1.17.1
html5lib==1.1
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.11
MarkupSafe==1.1.1
packaging==20.4
Pillow==7.2.0
pycparser==2.20
pyparsing==2.4.7
Pyphen==0.9.5
pytz==2020.1
reportlab==3.5.48
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
